Updated the High Level DynamoDB examples
