--- dotnetv3/dynamodb/HighLevelItemCRUD/HighLevelItemCRUD.cs
@@ -1,87 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.HighLevelItemCRUD]
-using System;
-using System.Collections.Generic;
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-
-namespace HighLevelItemCRUD
-{
-    [DynamoDBTable("ProductCatalog")]
-    public class Book
-    {
-        [DynamoDBHashKey] //Partition key
-        public int Id
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public string Title
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public string Isbn
-        {
-            get; set;
-        }
-        [DynamoDBProperty("Authors")] //String Set datatype
-        public List<string> BookAuthors
-        {
-            get; set;
-        }
-    }
-
-    public class HighLevelItemCrud
-    {
-        public static async void TestCrudOperations(IDynamoDBContext context)
-        {
-            int bookId = 1001; // Some unique value.
-            Book myBook = new Book
-            {
-                Id = bookId,
-                Title = "object persistence-AWS SDK for.NET SDK-Book 1001",
-                Isbn = "111-1111111001",
-                BookAuthors = new List<string> { "Author 1", "Author 2" },
-            };
-
-            // Save the book.
-            await context.SaveAsync(myBook);
-            // Retrieve the book.
-            Book bookRetrieved = await context.LoadAsync<Book>(bookId);
-
-            // Update few properties.
-            bookRetrieved.Isbn = "222-2222221001";
-            bookRetrieved.BookAuthors = new List<string> { " Author 1", "Author x" }; // Replace existing authors list with this.
-            await context.SaveAsync(bookRetrieved);
-
-            // Retrieve the updated book. This time add the optional ConsistentRead parameter using DynamoDBContextConfig object.
-            await context.LoadAsync<Book>(bookId, new DynamoDBContextConfig
-            {
-                ConsistentRead = true
-            });
-
-            // Delete the book.
-            await context.DeleteAsync<Book>(bookId);
-
-            // Try to retrieve deleted book. It should return null.
-            Book deletedBook = await context.LoadAsync<Book>(bookId, new DynamoDBContextConfig
-            {
-                ConsistentRead = true
-            });
-
-            if (deletedBook == null)
-            {
-                Console.WriteLine("Book is deleted");
-            }
-        }
-        static void Main()
-        {
-            var client = new AmazonDynamoDBClient();
-            DynamoDBContext context = new DynamoDBContext(client);
-            TestCrudOperations(context);
-        }
-    }
-}
-// snippet-end:[dynamodb.dotnet35.HighLevelItemCRUD]
