--- dotnetv3/dynamodb/HighLevelMappingArbitraryData/HighLevelMappingArbitraryData.cs
@@ -1,139 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.HighLevelMappingArbitraryData]
-using System;
-using System.Collections.Generic;
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.DynamoDBv2.DocumentModel;
-
-namespace HighLevelMappingArbitraryData
-{
-    [DynamoDBTable("ProductCatalog")]
-    public class Book
-    {
-        [DynamoDBHashKey] //Partition key
-        public int Id
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public string Title
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public string Isbn
-        {
-            get; set;
-        }
-        // Multi-valued (set type) attribute.
-        [DynamoDBProperty("Authors")]
-        public List<string> BookAuthors
-        {
-            get; set;
-        }
-        // Arbitrary type, with a converter to map it to DynamoDB type.
-        [DynamoDBProperty(typeof(DimensionTypeConverter))]
-        public DimensionType Dimensions
-        {
-            get; set;
-        }
-    }
-
-    public class DimensionType
-    {
-        public decimal Length
-        {
-            get; set;
-        }
-        public decimal Height
-        {
-            get; set;
-        }
-        public decimal Thickness
-        {
-            get; set;
-        }
-    }
-
-    // Converts the complex type DimensionType to string and vice-versa.
-    public class DimensionTypeConverter : IPropertyConverter
-    {
-        public DynamoDBEntry ToEntry(object value)
-        {
-            DimensionType bookDimensions = value as DimensionType;
-            if (bookDimensions == null) throw new ArgumentOutOfRangeException();
-
-            string data = string.Format("{1}{0}{2}{0}{3}", " x ",
-                            bookDimensions.Length, bookDimensions.Height, bookDimensions.Thickness);
-
-            DynamoDBEntry entry = new Primitive
-            {
-                Value = data
-            };
-            return entry;
-        }
-
-        public object FromEntry(DynamoDBEntry entry)
-        {
-            Primitive primitive = entry as Primitive;
-            if (primitive == null || !(primitive.Value is String) || string.IsNullOrEmpty((string)primitive.Value))
-                throw new ArgumentOutOfRangeException();
-
-            string[] data = ((string)(primitive.Value)).Split(new[] { " x " }, StringSplitOptions.None);
-            if (data.Length != 3) throw new ArgumentOutOfRangeException();
-
-            DimensionType complexData = new DimensionType
-            {
-                Length = Convert.ToDecimal(data[0]),
-                Height = Convert.ToDecimal(data[1]),
-                Thickness = Convert.ToDecimal(data[2])
-            };
-            return complexData;
-        }
-    }
-    public class HighLevelMappingArbitraryData
-    {
-        public static async void AddRetrieveUpdateBook(IDynamoDBContext context)
-        {
-            // Create a book.
-            DimensionType myBookDimensions = new DimensionType()
-            {
-                Length = 8M,
-                Height = 11M,
-                Thickness = 0.5M
-            };
-
-            Book myBook = new Book
-            {
-                Id = 501,
-                Title = "AWS SDK for .NET Object Persistence Model Handling Arbitrary Data",
-                Isbn = "999-9999999999",
-                BookAuthors = new List<string> { "Author 1", "Author 2" },
-                Dimensions = myBookDimensions
-            };
-
-            await context.SaveAsync(myBook);
-
-            // Retrieve the book.
-            Book bookRetrieved = await context.LoadAsync<Book>(501);
-
-            // 3. Update property (book dimensions).
-            bookRetrieved.Dimensions.Height += 1;
-            bookRetrieved.Dimensions.Length += 1;
-            bookRetrieved.Dimensions.Thickness += 0.2M;
-
-            // Update the book.
-            await context.SaveAsync(bookRetrieved);
-        }
-
-        static void Main()
-        {
-            var client = new AmazonDynamoDBClient();
-            DynamoDBContext context = new DynamoDBContext(client);
-            AddRetrieveUpdateBook(context);            
-        }
-    }
-}
-// snippet-end:[dynamodb.dotnet35.HighLevelMappingArbitraryData]
