--- dotnetv3/dynamodb/HighLevelQueryAndScan/HighLevelQueryAndScan.cs
@@ -1,279 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.HighLevelQueryAndScan]
-using System;
-using System.Collections.Generic;
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.DynamoDBv2.DocumentModel;
-
-namespace HighLevelQueryAndScan
-{
-    
-    [DynamoDBTable("Reply")]
-    public class Reply
-    {
-        [DynamoDBHashKey] //Partition key
-        public string Id
-        {
-            get; set;
-        }
-
-        [DynamoDBRangeKey] //Sort key
-        public DateTime ReplyDateTime
-        {
-            get; set;
-        }
-
-        // Properties included implicitly.
-        public string Message
-        {
-            get; set;
-        }
-        // Explicit property mapping with object persistence model attributes.
-        [DynamoDBProperty("LastPostedBy")]
-        public string PostedBy
-        {
-            get; set;
-        }
-        // Property to store version number for optimistic locking.
-        [DynamoDBVersion]
-        public int? Version
-        {
-            get; set;
-        }
-    }
-
-    [DynamoDBTable("Thread")]
-    public class Thread
-    {
-        // Partition key mapping.
-        [DynamoDBHashKey] //Partition key
-        public string ForumName
-        {
-            get; set;
-        }
-        [DynamoDBRangeKey] //Sort key
-        public DateTime Subject
-        {
-            get; set;
-        }
-        // Implicit mapping.
-        public string Message
-        {
-            get; set;
-        }
-        public string LastPostedBy
-        {
-            get; set;
-        }
-        public int Views
-        {
-            get; set;
-        }
-        public int Replies
-        {
-            get; set;
-        }
-        public bool Answered
-        {
-            get; set;
-        }
-        public DateTime LastPostedDateTime
-        {
-            get; set;
-        }
-        // Explicit mapping (property and table attribute names are different).
-        [DynamoDBProperty("Tags")]
-        public List<string> KeywordTags
-        {
-            get; set;
-        }
-        // Property to store version number for optimistic locking.
-        [DynamoDBVersion]
-        public int? Version
-        {
-            get; set;
-        }
-    }
-
-    [DynamoDBTable("Forum")]
-    public class Forum
-    {
-        [DynamoDBHashKey]
-        public string Name
-        {
-            get; set;
-        }
-        // All the following properties are explicitly mapped
-        // to show how to provide mapping.
-        [DynamoDBProperty]
-        public int Threads
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public int Views
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public string LastPostBy
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public DateTime LastPostDateTime
-        {
-            get; set;
-        }
-        [DynamoDBProperty]
-        public int Messages
-        {
-            get; set;
-        }
-    }
-
-    [DynamoDBTable("ProductCatalog")]
-    public class Book
-    {
-        [DynamoDBHashKey] //Partition key
-        public int Id
-        {
-            get; set;
-        }
-        public string Title
-        {
-            get; set;
-        }
-        public string Isbn
-        {
-            get; set;
-        }
-        public int Price
-        {
-            get; set;
-        }
-        public string PageCount
-        {
-            get; set;
-        }
-        public string ProductCategory
-        {
-            get; set;
-        }
-        public bool InPublication
-        {
-            get; set;
-        }
-    }
-
-    public class HighLevelQueryAndScan
-    {        
-        public static async void GetBook(IDynamoDBContext context, int productId)
-        {
-            Book bookItem = await context.LoadAsync<Book>(productId);
-
-            Console.WriteLine("\nGetBook: Printing result.....");
-            Console.WriteLine("Title: {0} \n No.Of threads:{1} \n No. of messages: {2}",
-                      bookItem.Title, bookItem.Isbn, bookItem.PageCount);
-        }
-
-        public static async void FindRepliesInLast15Days(IDynamoDBContext context,
-                                string forumName,
-                                string threadSubject)
-        {
-            string replyId = forumName + "#" + threadSubject;
-            DateTime twoWeeksAgoDate = DateTime.UtcNow - TimeSpan.FromDays(15);
-            
-            List<object> times = new List<object>();
-            times.Add(twoWeeksAgoDate);            
-
-            List<ScanCondition> scs = new List<ScanCondition>();
-            var sc = new ScanCondition("LastPostedBy", ScanOperator.GreaterThan, times.ToArray());
-            scs.Add(sc);
-
-            var cfg = new DynamoDBOperationConfig {
-                QueryFilter = scs
-            };
-
-            AsyncSearch<Reply> response = context.QueryAsync<Reply>(replyId, cfg);
-            IEnumerable<Reply> latestReplies = await response.GetRemainingAsync();
-
-            Console.WriteLine("\nReplies in last 15 days:");
-
-            foreach (Reply r in latestReplies)
-                Console.WriteLine("{0}\t{1}\t{2}\t{3}", r.Id, r.PostedBy, r.Message, r.ReplyDateTime);
-        }
-
-        public static async void FindRepliesPostedWithinTimePeriod(IDynamoDBContext context,
-                                      string forumName,
-                                      string threadSubject)
-        {
-            string forumId = forumName + "#" + threadSubject;
-            Console.WriteLine("\nReplies posted within time period:");
-
-            DateTime startDate = DateTime.UtcNow - TimeSpan.FromDays(30);
-            DateTime endDate = DateTime.UtcNow - TimeSpan.FromDays(1);
-
-            List<object> times = new List<object>();
-            times.Add(startDate);
-            times.Add(endDate);
-
-            List<ScanCondition> scs = new List<ScanCondition>();
-            var sc = new ScanCondition("LastPostedBy", ScanOperator.Between, times.ToArray());
-            scs.Add(sc);
-
-            var cfg = new DynamoDBOperationConfig
-            {
-                QueryFilter = scs
-            };
-
-            AsyncSearch<Reply> response = context.QueryAsync<Reply>(forumId, cfg);
-            IEnumerable<Reply> repliesInAPeriod = await response.GetRemainingAsync();
-
-            foreach (Reply r in repliesInAPeriod)
-                Console.WriteLine("{0}\t{1}\t{2}\t{3}", r.Id, r.PostedBy, r.Message, r.ReplyDateTime);
-        }
-
-        public static async void FindProductsPricedLessThanZero(IDynamoDBContext context)
-        {
-            int price = 0;
-
-            List<ScanCondition> scs = new List<ScanCondition>();
-            var sc1 = new ScanCondition("Price", ScanOperator.LessThan, price);
-            var sc2 = new ScanCondition("ProductCategory", ScanOperator.Equal, "Book");
-            scs.Add(sc1);
-            scs.Add(sc2);
-
-            AsyncSearch<Book> response = context.ScanAsync<Book>(scs);
-
-            IEnumerable<Book> itemsWithWrongPrice = await response.GetRemainingAsync();
-                
-            Console.WriteLine("\nFindProductsPricedLessThanZero: Printing result.....");
-
-            foreach (Book r in itemsWithWrongPrice)
-                Console.WriteLine("{0}\t{1}\t{2}\t{3}", r.Id, r.Title, r.Price, r.Isbn);
-        }
-
-        static void Main(string[] args)
-        {
-            var client = new AmazonDynamoDBClient();
-
-            DynamoDBContext context = new DynamoDBContext(client);
-            // Get an item.
-            GetBook(context, 101);
-
-            // Sample forum and thread to test queries.
-            string forumName = "Amazon DynamoDB";
-            string threadSubject = "DynamoDB Thread 1";
-            // Sample queries.
-            FindRepliesInLast15Days(context, forumName, threadSubject);
-            FindRepliesPostedWithinTimePeriod(context, forumName, threadSubject);
-
-            // Scan table.
-            FindProductsPricedLessThanZero(context);
-        }
-    }
-}
-// snippet-end:[dynamodb.dotnet35.HighLevelQueryAndScan]
